# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Financial Evidence Lab API
  version: 0.9.0
  description: >-
    Frozen v1 contract surface (ADR-0001). This file seeds the M0 freeze:
    conventions (error envelope, idempotency, optimistic concurrency,
    run-ID pattern), identity/workspace resources, evidence reads, and
    ingestion jobs. v0.3.0 adds observable hybrid retrieval (ADR-0006).
    v0.4.0 adds bounded agentic extraction (ADR-0007): extraction runs,
    typed resumable SSE, proposals/review, approved versions, plus a
    compatible open-enum loosening for volatile M2 planner/lane/event
    labels. v0.5.0 declares the API hardening surface (#191): createQuery
    and createQueryRerun answer 402 COST_LIMIT_EXCEEDED when a user daily
    or organization monthly hard cost ceiling (spec 18.2) would be crossed,
    and those two plus createRetrievalFeedback answer 429 RATE_LIMITED with
    Retry-After when the per-organization token bucket for the route is
    exhausted. Both are additive response codes on existing operations; no
    path, parameter, field, or enum changes. List bounding and its query
    parameter are deliberately not declared here — lists stay unbounded
    until pagination lands with a client that can page. Modeling and
    forecasting paths arrive in later minors.
servers:
  - url: /
# `x-fel-status: planned` on a path item means the contract is frozen but the
# FastAPI service does not serve that path yet. apps/api/tests/test_openapi_parity.py
...
paths:
...
  /v1/workspaces/{workspaceId}/extractions:
    get:
      operationId: listExtractions
      summary: List extraction proposals for a workspace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: state
          in: query
          schema:
            $ref: "#/components/schemas/ProposalState"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageOrder"
      responses:
        "200":
          description: Extraction proposals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionProposalPage"
        default:
          $ref: "#/components/responses/Error"
        "409":
          description:
            "CONFLICT: incompatible merge, terminal action, contradictory resolved alternative
            or IDEMPOTENCY_KEY_REUSED."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description:
            "PRECONDITION_FAILED: selected proposal, approved head or complete conflict membership/ETag
            changed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description:
            "EXTRACTION_TOO_LARGE: bounded request/evidence/group/event ceiling; safe resource
            details, no partial validation."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description:
            "VALIDATION_ERROR: malformed closed input/cursor, blocked financial/citation/temporal
            validation or unavailable validation policy."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      description:
        Explicit bounded page. Cursor binds endpoint, tenant, resource, cutoff/corpus and filters;
        reauthorize every page. Deterministic before/after order is not a cross-request MVCC snapshot.
  /v1/extractions/{extractionId}:
    parameters:
      - $ref: "#/components/parameters/ExtractionId"
    get:
      operationId: getExtraction
      summary: Read one extraction proposal
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Extraction proposal.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionProposal"
        default:
          $ref: "#/components/responses/Error"
        "409":
          description:
            "CONFLICT: incompatible merge, terminal action, contradictory resolved alternative
            or IDEMPOTENCY_KEY_REUSED."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description:
            "PRECONDITION_FAILED: selected proposal, approved head or complete conflict membership/ETag
            changed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description:
            "EXTRACTION_TOO_LARGE: bounded request/evidence/group/event ceiling; safe resource
            details, no partial validation."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description:
            "VALIDATION_ERROR: malformed closed input/cursor, blocked financial/citation/temporal
            validation or unavailable validation policy."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/approved-extractions/{recordId}:
    get:
      operationId: getApprovedExtraction
      summary: Read the current approved extraction version
      security:
        - bearerAuth: []
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Current approved version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovedExtraction"
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        default:
          $ref: "#/components/responses/Error"
        "409":
          description:
            "CONFLICT: incompatible merge, terminal action, contradictory resolved alternative
            or IDEMPOTENCY_KEY_REUSED."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description:
            "PRECONDITION_FAILED: selected proposal, approved head or complete conflict membership/ETag
            changed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description:
            "EXTRACTION_TOO_LARGE: bounded request/evidence/group/event ceiling; safe resource
            details, no partial validation."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description:
            "VALIDATION_ERROR: malformed closed input/cursor, blocked financial/citation/temporal
            validation or unavailable validation policy."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/approved-extractions/{recordId}/versions:
    get:
      operationId: listApprovedExtractionVersions
      summary: List immutable approved extraction versions
      security:
        - bearerAuth: []
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageOrder"
      responses:
        "200":
          description: Immutable history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovedExtractionPage"
        default:
          $ref: "#/components/responses/Error"
        "409":
          description:
            "CONFLICT: incompatible merge, terminal action, contradictory resolved alternative
            or IDEMPOTENCY_KEY_REUSED."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description:
            "PRECONDITION_FAILED: selected proposal, approved head or complete conflict membership/ETag
            changed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description:
            "EXTRACTION_TOO_LARGE: bounded request/evidence/group/event ceiling; safe resource
            details, no partial validation."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description:
            "VALIDATION_ERROR: malformed closed input/cursor, blocked financial/citation/temporal
            validation or unavailable validation policy."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      description:
        Explicit bounded page. Cursor binds endpoint, tenant, resource, cutoff/corpus and filters;
        reauthorize every page. Deterministic before/after order is not a cross-request MVCC snapshot.
...
  /v1/extraction-runs/{runId}/steps:
    get:
      operationId: listExtractionSteps
      summary: listExtractionSteps
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageOrder"
      responses:
        "200":
          description: Bounded extraction resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionStepPage"
        default:
          $ref: "#/components/responses/Error"
        "409":
          description:
            "CONFLICT: incompatible merge, terminal action, contradictory resolved alternative
            or IDEMPOTENCY_KEY_REUSED."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description:
            "PRECONDITION_FAILED: selected proposal, approved head or complete conflict membership/ETag
            changed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description:
            "EXTRACTION_TOO_LARGE: bounded request/evidence/group/event ceiling; safe resource
            details, no partial validation."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description:
            "VALIDATION_ERROR: malformed closed input/cursor, blocked financial/citation/temporal
            validation or unavailable validation policy."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/extraction-runs/{runId}/event-history:
    get:
      operationId: listExtractionEvents
      summary: listExtractionEvents
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageOrder"
      responses:
        "200":
          description: Bounded extraction resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionEventPage"
        default:
          $ref: "#/components/responses/Error"
        "409":
          description:
            "CONFLICT: incompatible merge, terminal action, contradictory resolved alternative
            or IDEMPOTENCY_KEY_REUSED."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description:
            "PRECONDITION_FAILED: selected proposal, approved head or complete conflict membership/ETag
            changed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description:
            "EXTRACTION_TOO_LARGE: bounded request/evidence/group/event ceiling; safe resource
            details, no partial validation."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description:
            "VALIDATION_ERROR: malformed closed input/cursor, blocked financial/citation/temporal
            validation or unavailable validation policy."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/workspaces/{workspaceId}/extraction-conflicts:
    get:
      operationId: listExtractionConflicts
      summary: listExtractionConflicts
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: status
          in: query
          schema:
            enum:
              - open
              - resolved
              - superseded
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageOrder"
      responses:
        "200":
          description: Bounded extraction resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionConflictPage"
        default:
          $ref: "#/components/responses/Error"
        "409":
          description:
            "CONFLICT: incompatible merge, terminal action, contradictory resolved alternative
            or IDEMPOTENCY_KEY_REUSED."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description:
            "PRECONDITION_FAILED: selected proposal, approved head or complete conflict membership/ETag
            changed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description:
            "EXTRACTION_TOO_LARGE: bounded request/evidence/group/event ceiling; safe resource
            details, no partial validation."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description:
            "VALIDATION_ERROR: malformed closed input/cursor, blocked financial/citation/temporal
            validation or unavailable validation policy."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/extraction-conflicts/{conflictId}:
    get:
      operationId: getExtractionConflict
      summary: getExtractionConflict
      security:
        - bearerAuth: []
      parameters:
        - name: conflictId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Bounded extraction resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionConflict"
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        default:
          $ref: "#/components/responses/Error"
        "409":
          description:
            "CONFLICT: incompatible merge, terminal action, contradictory resolved alternative
            or IDEMPOTENCY_KEY_REUSED."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description:
            "PRECONDITION_FAILED: selected proposal, approved head or complete conflict membership/ETag
            changed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description:
            "EXTRACTION_TOO_LARGE: bounded request/evidence/group/event ceiling; safe resource
            details, no partial validation."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description:
            "VALIDATION_ERROR: malformed closed input/cursor, blocked financial/citation/temporal
            validation or unavailable validation policy."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/approved-extractions/{recordId}/versions/{versionId}:
    get:
      operationId: getApprovedExtractionVersion
      summary: getApprovedExtractionVersion
      security:
        - bearerAuth: []
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: versionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Bounded extraction resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovedExtraction"
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        default:
          $ref: "#/components/responses/Error"
        "409":
          description:
            "CONFLICT: incompatible merge, terminal action, contradictory resolved alternative
            or IDEMPOTENCY_KEY_REUSED."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description:
            "PRECONDITION_FAILED: selected proposal, approved head or complete conflict membership/ETag
            changed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description:
            "EXTRACTION_TOO_LARGE: bounded request/evidence/group/event ceiling; safe resource
            details, no partial validation."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description:
            "VALIDATION_ERROR: malformed closed input/cursor, blocked financial/citation/temporal
            validation or unavailable validation policy."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Authenticated tenant context (ADR-0004 read path); org membership
        is enforced server-side on every corpus read.
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ExtractionId:
      name: extractionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageCursor:
      name: cursor
      in: query
      description: Opaque scope-bound continuation; reauthenticate every request. Not a cross-request MVCC snapshot.
      schema: { type: string, maxLength: 2048 }
    PageOrder:
      name: order
      in: query
      schema: { type: string, enum: [asc, desc], default: asc }
...
  headers:
...
    ETag:
      description: Version tag for optimistic concurrency.
      schema:
        type: string
  responses:
...
    Error:
      description: Error envelope (all non-2xx responses use this shape).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: Stable machine-readable code, e.g. TEMPORAL_SCOPE_VIOLATION.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
...
    ProposalState:
      type: string
      enum:
        - proposed
        - needs_review
        - accepted
        - rejected
        - superseded
      x-fel-open-enum: false
...
    ExtractionCandidateFields:
      $ref: ../schemas/extraction-candidate-fields.schema.json
    ExtractionPayload:
      $ref: ../schemas/extraction-payload.schema.json
    ExtractionProposal:
      type: object
      additionalProperties: false
      required:
        - id
        - run_id
        - kind
        - metric_id
        - payload
        - evidence
        - record_confidence
        - field_confidences
        - validations
        - state
        - review_priority
        - version
        - conflict_ids
      properties:
        id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
            - kpi
            - guidance
            - revenue_driver
        metric_id:
          type: string
        payload:
          oneOf:
            - $ref: "#/components/schemas/ExtractionPayload"
            - $ref: "#/components/schemas/ExtractionCandidateFields"
          description: >-
            Ordinary schema-valid, display-safe payloads keep their strict shape. Otherwise return
            the read-only candidate fields projection, including for fractional or unsafe JSON
            numeric leaves (recursively excluding booleans and numeric strings). This alternative
            implies no financial validity or approval. Preflight stored payload bytes at 1 MiB;
            overflow returns a typed 413 with the safe resource ID, never a truncated candidate.
        evidence:
          type: array
          minItems: 0
          items:
            $ref: "#/components/schemas/EvidenceEdge"
        record_confidence:
          type:
            - string
            - "null"
          pattern: ^(0(\.\d{1,3})?|1(\.0{1,3})?)$
          description: Null means uncalibrated; never convert to zero or invent a score.
        field_confidences:
          type: object
          additionalProperties:
            type: string
        validations:
          type: array
          items:
            type: object
            required:
              - code
              - status
            properties:
              code:
                type: string
              status:
                type: string
                enum:
                  - pass
                  - warning
                  - fail
              detail:
                type: object
                additionalProperties: true
        state:
          $ref: "#/components/schemas/ProposalState"
        review_priority:
          type: string
          enum:
            - normal
            - high
        version:
          type: integer
          minimum: 1
        conflict_ids:
          type: array
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            format: uuid
    EvidenceEdge:
      type: object
      additionalProperties: false
      required:
        - source_span_id
        - document_version_id
        - role
        - citation_status
      properties:
        source_span_id:
          type: string
          format: uuid
        document_version_id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - supports
            - definition
            - conflicts
            - derivation_input
        citation_status:
          type: string
          enum:
            - verified
            - partial
            - contradictory
            - invalid
...
    ApprovedExtraction:
      type: object
      additionalProperties: false
      required:
        - record_id
        - version_id
        - version
        - kind
        - metric_id
        - payload
        - evidence
        - evidence_manifest_hash
        - ontology_version
        - approved_by
        - created_at
        - approval_reason
        - normalizer_version
        - validator_version
        - validation_context
      properties:
        record_id:
          type: string
          format: uuid
        version_id:
          type: string
          format: uuid
        parent_version_id:
          type:
            - string
            - "null"
          format: uuid
        version:
          type: integer
          minimum: 1
        kind:
          type: string
          enum:
            - kpi
            - guidance
            - revenue_driver
        metric_id:
          type: string
        payload:
          $ref: "#/components/schemas/ExtractionPayload"
        evidence:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/EvidenceEdge"
        evidence_manifest_hash:
          type: string
          pattern: ^sha256:[0-9a-f]{64}$
        ontology_version:
          type: string
        approved_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        approval_reason:
          type: string
        normalizer_version:
          type: string
        validator_version:
          type: string
        validation_context:
          anyOf:
            - $ref: "#/components/schemas/ExtractionValidationContext"
            - type: "null"
          description: Historical null is unknown provenance; new approved versions require complete context.
    ExtractionConflict:
      $ref: ../schemas/extraction-conflict.schema.json
    ExtractionValidationContext:
      $ref: ../schemas/extraction-validation-context.schema.json
    ExtractionEventPage:
      $ref: ../schemas/extraction-event-page.schema.json
    ExtractionStepSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - step_name
        - attempt
        - status
        - input_hash
        - output_hash
        - started_at
        - finished_at
      properties:
        id:
          type: string
          format: uuid
        step_name:
          type: string
        attempt:
          type: integer
          minimum: 1
        status:
          enum:
            - pending
            - running
            - succeeded
            - failed
            - skipped
            - cancelled
        input_hash:
          type: string
        output_hash:
          type:
            - string
            - "null"
        started_at:
          type:
            - string
            - "null"
          format: date-time
        finished_at:
          type:
            - string
            - "null"
          format: date-time
        input_tokens:
          type: integer
          minimum: 0
        output_tokens:
          type: integer
          minimum: 0
        cost_usd:
          type: string
        error:
          anyOf:
            - $ref: "#/components/schemas/Error"
            - type: "null"
...
    ExtractionProposalPage:
      type: object
      additionalProperties: false
      required:
        - items
        - next_cursor
        - previous_cursor
        - limit
      properties:
        items:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/ExtractionProposal"
        next_cursor:
          type:
            - string
            - "null"
          maxLength: 2048
        previous_cursor:
          type:
            - string
            - "null"
          maxLength: 2048
        limit:
          type: integer
          minimum: 1
          maximum: 200
    ApprovedExtractionPage:
      type: object
      additionalProperties: false
      required:
        - items
        - next_cursor
        - previous_cursor
        - limit
      properties:
        items:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/ApprovedExtraction"
        next_cursor:
          type:
            - string
            - "null"
          maxLength: 2048
        previous_cursor:
          type:
            - string
            - "null"
          maxLength: 2048
        limit:
          type: integer
          minimum: 1
          maximum: 200
    ExtractionConflictPage:
      type: object
      additionalProperties: false
      required:
        - items
        - next_cursor
        - previous_cursor
        - limit
      properties:
        items:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/ExtractionConflict"
        next_cursor:
          type:
            - string
            - "null"
          maxLength: 2048
        previous_cursor:
          type:
            - string
            - "null"
          maxLength: 2048
        limit:
          type: integer
          minimum: 1
          maximum: 200
    ExtractionStepPage:
      type: object
      additionalProperties: false
      required:
        - items
        - next_cursor
        - previous_cursor
        - limit
      properties:
        items:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/ExtractionStepSummary"
        next_cursor:
          type:
            - string
            - "null"
          maxLength: 2048
        previous_cursor:
          type:
            - string
            - "null"
          maxLength: 2048
        limit:
          type: integer
          minimum: 1
          maximum: 200